move and parametrize test
